Remove code duplications.
