+++ src/main/java/com/minecolonies/coremod/client/gui/AbstractWindowBuilding.java
@@ -2,12 +2,14 @@ package com.minecolonies.coremod.client.gui;
 
 import com.minecolonies.blockout.controls.Button;
 import com.minecolonies.blockout.controls.Label;
+import com.minecolonies.blockout.views.SwitchView;
 import com.minecolonies.coremod.MineColonies;
 import com.minecolonies.coremod.colony.buildings.AbstractBuildingHut;
 import com.minecolonies.coremod.colony.buildings.AbstractBuildingWorker;
 import com.minecolonies.coremod.network.messages.BuildRequestMessage;
 import com.minecolonies.coremod.network.messages.OpenInventoryMessage;
 import com.minecolonies.coremod.util.LanguageHandler;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Manage windows associated with Buildings.
@@ -20,6 +22,12 @@ public abstract class AbstractWindowBuilding<B extends AbstractBuildingHut.View>
     private static final String BUTTON_REPAIR       = "repair";
     private static final String BUTTON_INVENTORY    = "inventory";
     private static final String LABEL_BUILDING_NAME = "name";
+    private static final String BUTTON_PREVPAGE             = "prevPage";
+    private static final String BUTTON_NEXTPAGE             = "nextPage";
+    private static final String VIEW_PAGES = "pages";
+
+    private final Button buttonPrevPage;
+    private final Button buttonNextPage;
 
     /**
      * Type B is a class that extends {@link AbstractBuildingWorker.View}.
@@ -40,6 +48,8 @@ public abstract class AbstractWindowBuilding<B extends AbstractBuildingHut.View>
         registerButton(BUTTON_BUILD, this::buildClicked);
         registerButton(BUTTON_REPAIR, this::repairClicked);
         registerButton(BUTTON_INVENTORY, this::inventoryClicked);
+        buttonNextPage = findPaneOfTypeByID(BUTTON_NEXTPAGE, Button.class);
+        buttonPrevPage = findPaneOfTypeByID(BUTTON_PREVPAGE, Button.class);
     }
 
     /**
@@ -72,6 +82,10 @@ public abstract class AbstractWindowBuilding<B extends AbstractBuildingHut.View>
     @Override
     public void onOpened()
     {
+        if(buttonPrevPage != null)
+        {
+            findPaneOfTypeByID(BUTTON_PREVPAGE, Button.class).setEnabled(false);
+        }
         findPaneOfTypeByID(LABEL_BUILDING_NAME, Label.class).setLabelText(LanguageHandler.getString(getBuildingName()) + " " + building.getBuildingLevel());
 
         if (building.getBuildingLevel() == 0)
@@ -84,6 +98,27 @@ public abstract class AbstractWindowBuilding<B extends AbstractBuildingHut.View>
             final Button button = findPaneOfTypeByID(BUTTON_BUILD, Button.class);
             button.setLabel(LanguageHandler.getString("com.minecolonies.coremod.gui.workerHuts.upgradeUnavailable"));
             button.disable();
+        }
+    }
+
+    @Override
+    public void onButtonClicked(@NotNull final Button button)
+    {
+        switch (button.getID())
+        {
+            case BUTTON_PREVPAGE:
+                findPaneOfTypeByID(VIEW_PAGES, SwitchView.class).previousView();
+                buttonPrevPage.setEnabled(false);
+                buttonNextPage.setEnabled(true);
+                break;
+            case BUTTON_NEXTPAGE:
+                findPaneOfTypeByID(VIEW_PAGES, SwitchView.class).nextView();
+                buttonPrevPage.setEnabled(true);
+                buttonNextPage.setEnabled(false);
+                break;
+            default:
+                super.onButtonClicked(button);
+                break;
         }
     }
