+++ src/main/java/com/minecolonies/coremod/client/gui/WindowHutDeliveryman.java
@@ -11,14 +11,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public class WindowHutDeliveryman extends AbstractWindowWorkerBuilding<BuildingDeliveryman.View> implements Button.Handler
 {
-    private static final String BUTTON_PREV_PAGE = "prevPage";
-    private static final String BUTTON_NEXT_PAGE = "nextPage";
-    private static final String VIEW_PAGES       = "pages";
-
     private static final String HUT_WAREHOUSE_RESOURCE_SUFFIX = ":gui/windowHutWarehouse.xml";
-
-    private Button buttonPrevPage;
-    private Button buttonNextPage;
 
     /**
      * Constructor for the window of the warehouse hut.
@@ -28,8 +21,6 @@ public class WindowHutDeliveryman extends AbstractWindowWorkerBuilding<BuildingD
     public WindowHutDeliveryman(@NotNull final BuildingDeliveryman.View building)
     {
         super(building, Constants.MOD_ID + HUT_WAREHOUSE_RESOURCE_SUFFIX);
-        registerButton(BUTTON_PREV_PAGE, this::prevClicked);
-        registerButton(BUTTON_NEXT_PAGE, this::nextClicked);
     }
 
     @NotNull
@@ -39,34 +30,4 @@ public class WindowHutDeliveryman extends AbstractWindowWorkerBuilding<BuildingD
         return "com.minecolonies.coremod.gui.workerHuts.warehouse";
     }
 
-    @Override
-    public void onOpened()
-    {
-        super.onOpened();
-
-        findPaneOfTypeByID(BUTTON_PREV_PAGE, Button.class).setEnabled(false);
-
-        buttonPrevPage = findPaneOfTypeByID(BUTTON_PREV_PAGE, Button.class);
-        buttonNextPage = findPaneOfTypeByID(BUTTON_NEXT_PAGE, Button.class);
-    }
-
-    /**
-     * Action performed when previous button is clicked.
-     */
-    private void prevClicked()
-    {
-        findPaneOfTypeByID(VIEW_PAGES, SwitchView.class).previousView();
-        buttonPrevPage.setEnabled(false);
-        buttonNextPage.setEnabled(true);
-    }
-
-    /**
-     * Action performed when next button is clicked.
-     */
-    private void nextClicked()
-    {
-        findPaneOfTypeByID(VIEW_PAGES, SwitchView.class).nextView();
-        buttonPrevPage.setEnabled(true);
-        buttonNextPage.setEnabled(false);
-    }
 }
